% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunomars_tweet.R
\docType{data}
\name{brunomars_tweet}
\alias{brunomars_tweet}
\title{Tweets data of 500 tweets from BrunoMars.}
\format{
A data frame with 500 rows and 2 variables:
\describe{
\item{time}{datetime, the time a tweet is sent}
\item{tweet}{text of the tweet}
...
}
}
\source{
\url{https://twitter.com/BrunoMars/}
}
\usage{
brunomars_tweet
}
\description{
A dataset containing the sent time and tweet text for 500
tweets from Twitter user BrunoMars.
}
\keyword{datasets}
