% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_score}
\alias{get_score}
\title{Get topics and, influence authors scores based on a search term or a hashtag.}
\usage{
get_score(keyword)
}
\arguments{
\item{keyword}{Your target search-term or hashtag to get the influence score of similar topics and users.}
}
\value{
A dataframe.
}
\description{
Get topics and, influence authors scores based on a search term or a hashtag.
}
\examples{
get_score('#Bitcoin')
}
