% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_trends}
\alias{get_trends}
\title{Get trends}
\usage{
get_trends(id = "1")
}
\arguments{
\item{id}{Location id, Default to 1 \strong{worldwide}. Get list of available locations from \code{data('locID', package = 'tweetr')}.}
}
\value{
A dataframe.
}
\description{
Get trends
}
