% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{Get tweets data from custom search}
\usage{
get_tweets(
  query = NA,
  lat = NA,
  long = NA,
  radius = NA,
  place = NA,
  lang = NULL,
  since = NA,
  until = NA,
  from = NA,
  to = NA,
  replies = F,
  minLikes = NA,
  minReplies = NA,
  minRetweets = NA,
  verified = F,
  hasImage = F,
  hasVideo = F,
  hasMedia = F,
  hasLinks = F,
  url = NA,
  count = "-1"
)
}
\arguments{
\item{query}{Can be hashtag, keyword, or a user..}

\item{lat}{Latitude value to provide for geographical search (ex: \code{33.575692}); Latitude, longitude, and radius are each necessary parameters for geographical search.}

\item{long}{Longitude value to provide for geographical search (ex: \code{-7.625285}); Latitude, longitude, and radius are each necessary parameters for geographical search.}

\item{radius}{Match tweets with tweet-specific location information that are within a sphere-shaped geographic area. (ex: 100km); Latitude, longitude, and radius are each necessary parameters for geographical search.}

\item{place}{Match tweets nearby a city or country.}

\item{lang}{Filter tweets by language (ex: 'en'). Default to \code{NULL}}

\item{since}{Filter tweets from a specific date. Allowed date format: \code{YYYY/MM/DD}}

\item{until}{Filter tweets upto a specific date. Allowed date format: \code{YYYY/MM/DD}}

\item{from}{Collect tweets from a user. (ex: \verb{@BillGates})}

\item{to}{Collect tweets mentioning or replying to a specific user. (ex: \verb{@BillGates})}

\item{replies}{Set to false by default. if it's true then it will fetch tweets replies only.}

\item{minLikes}{Tweets with at least \code{(n)} likes should be filtered.}

\item{minReplies}{Tweets with at least \code{(n)} replies should be filtered.}

\item{minRetweets}{Tweets with at least \code{(n)} retweets should be filtered.}

\item{verified}{If true, only tweets from verified accounts will be filtered.}

\item{hasImage}{If true, only tweets with images should be filtered.}

\item{hasVideo}{If true, Only tweets with videos should be filtered.}

\item{hasMedia}{If true, Only tweets with medias should be filtered.}

\item{hasLinks}{If true, Only tweets with links should be filtered.}

\item{url}{Get tweets that only contain links to specific domain names (ex: Oscars.org); Alternatively, you may type in "oscars" to get all tweets with urls that mention the Oscars term.}

\item{count}{Default to \code{-1}, this will collect all tweets related to your search unless you specify a number of tweets to collect. Depending on the tweets volume, the process might take a while.}
}
\value{
A list.
}
\description{
Get tweets data from custom search
}
