% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetr.R
\name{plot_hashtags}
\alias{plot_hashtags}
\title{Plot hashtag word analysis}
\usage{
plot_hashtags(df, text_col)
}
\arguments{
\item{df}{data.frame}

\item{text_col}{A column name in data.frame}
}
\value{
A chart plotting analysis the most commonly used words.
}
\description{
A chart plotting analysis result of the most commonly used
words in tweets.
}
\examples{
plot_hashtags(tweet_data, time)
}
