% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetr.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Plot timeline analysis}
\usage{
plot_timeline(df, time)
}
\arguments{
\item{df}{data.frame}

\item{time}{A column name in data.frame}
}
\value{
A chart plotting the counts of tweets versus hours.
}
\description{
Analysis what time of day the tweets occurs and plot the
counts of tweets versus hours.
}
\examples{
tweet_data <- tweetr::brunomars_tweet
plot_timeline(tweet_data, time)

}
