% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetr.R
\name{sentiment_analysis}
\alias{sentiment_analysis}
\title{sentiment_analysis}
\usage{
sentiment_analysis(tweet)
}
\arguments{
\item{tweet}{data.frame}
}
\value{
tweet_senti data.frame
}
\description{
This function takes a tweet dataframe as input. The input dataframe should contain a column named 'tweet' that contains tweet text information.
The function analyzes the sentiment of each tweet and categorize them into 'positive', 'negative' and 'netrual'. The sentiment information will be added as a new
column to the input dataframe and saved as a new dataframe (output).
}
\examples{
sentiment_analysis(tweet_username_123)
}
