% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetr.R
\name{visualize_sentiments}
\alias{visualize_sentiments}
\title{Visualize Sentiment Analysis}
\usage{
visualize_sentiments(sentiment_df, plot_type = "Standard")
}
\arguments{
\item{sentiment_df}{A DataFrame containing sentiment column from sentiment_analysis}

\item{plot_type}{An optional string for which chart to return. Options are ("Standard", "Stacked", "Separate")}
}
\value{
A bar chart containing most common words in tweets, colour coded by sentiment.
}
\description{
Takes in the output of sentiment_analysis and creates
a visualization of user's tweets with sentimental analysis.
}
\examples{
visualize_sentiments(sentiment_df, plot_type = "Stacked")
}
