% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetr.R
\name{plot_hashtags}
\alias{plot_hashtags}
\title{Plot hashtag word analysis}
\usage{
plot_hashtags(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
A chart plotting analysis the most commonly used words.
}
\description{
A chart plotting analysis result of the most commonly used
words in tweets.
}
\examples{
tweet_data <- tweetr::brunomars_tweet
plot_hashtags(tweet_data)
}
